refactor: divide the code into components
