+++ src/Helper.ts
@@ -1,0 +1,9 @@
+import * as vscode from 'vscode';
+import type { Configuration } from './interfaces/Configuration';
+
+export function getConfiguration(): Configuration {
+  const config = vscode.workspace
+    .getConfiguration()
+    .get<Configuration>('commitizen');
+  return config!;
+}
