+++ src/interfaces/Configuration.ts
@@ -1,0 +1,7 @@
+export interface Configuration {
+  autoSync: boolean;
+  subjectLength: number;
+  showOutputChannel: 'off' | 'always' | 'onError';
+  quoteMessageInGitCommit: boolean;
+  capitalizeWindowsDriveLetter: boolean;
+}
