+++ src/interfaces/GitCmdArgs.ts
@@ -1,0 +1,4 @@
+export interface GitCmdArgs {
+  message: string;
+  cwd: string;
+}
